style(accounts payments): fix minor lint issues

+++ src-tauri/src/payments.rs
@@ -32,7 +32,7 @@ pub async fn pay_bolt11_invoice(bolt11: &str, nwc_uri: &str) -> Result<String, P
     let current_timestamp = SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap()
-        .as_secs() as u64;
+        .as_secs();
 
     if current_timestamp > invoice.expires_at().expect("invoice expiry time").as_secs() {
         return Err(PaymentError::ExpiredInvoice);
